Move BitConverter to shared CoreLib partition (dotnet/coreclr#12876)
